put string dtypes in the strings module
