+++ src/zarr/core/dtype/__init__.py
@@ -8,9 +8,7 @@ from zarr.core.dtype.npy.complex import Complex64, Complex128
 from zarr.core.dtype.npy.float import Float16, Float32, Float64
 from zarr.core.dtype.npy.int import Int8, Int16, Int32, Int64, UInt8, UInt16, UInt32, UInt64
 from zarr.core.dtype.npy.sized import (
-    FixedLengthASCII,
     FixedLengthBytes,
-    FixedLengthUTF32,
     Structured,
 )
 from zarr.core.dtype.npy.time import DateTime64, TimeDelta64
@@ -24,6 +22,8 @@ import numpy.typing as npt
 from zarr.core.common import JSON
 from zarr.core.dtype.npy.string import (
     _NUMPY_SUPPORTS_VLEN_STRING,
+    FixedLengthASCII,
+    FixedLengthUTF32,
     VariableLengthString,
 )
 from zarr.core.dtype.registry import DataTypeRegistry
